Remove SolanaClient and reorganize OpenSea
