+++ packages/common/src/models/index.ts
@@ -19,7 +19,6 @@ export * from './ImageSizes'
 export * from './Kind'
 export * from './Lineup'
 export * from './OnChain'
-export * from './OpenSea'
 export * from './OS'
 export * from './Playable'
 export * from './PlaylistLibrary'
