+++ packages/common/src/services/index.ts
@@ -8,8 +8,7 @@ export * from './wallet-client'
 export * from './env'
 export * from './explore'
 export * from './audio-player'
-export * from './opensea-client'
-export * from './solana-client'
+export * from './opensea'
 export * from './helius'
 export * from './collectibles-provider'
 export * from './track-download'
