--- packages/common/src/services/opensea-client/index.ts
@@ -1,2 +1,0 @@
-export * from './ethCollectibleHelpers'
-export * from './OpenSeaClient'
