+++ packages/common/src/services/opensea/index.ts
@@ -1,0 +1,1 @@
+export { OpenSeaClient } from './OpenSeaClient'
