--- packages/common/src/services/solana-client/index.ts
@@ -1,1 +1,0 @@
-export { SolanaClient } from './SolanaClient'
