+++ packages/web/src/app/web-player/WebPlayer.test.tsx
@@ -21,10 +21,6 @@ vitest.mock('store/backend/sagas', () => ({
   setupBackend: function* () {},
   waitForBackendSetup: () => true
 }))
-// Mock the solana collectibles client
-vitest.mock('services/solana-client/SolanaClient', () => ({
-  SolanaClient: () => {}
-}))
 vitest.mock('services/audius-sdk', async () => ({
   audiusSdk: {}
 }))
